attn_res: move the torch reference under the kimi_k3 folder
The read is a K3 architecture feature whose block size is already a field in
the K3 config, so both halves of it now live in one place. LICENSE-Kimi-K3
stays inside attn_res/, beside the only file it governs — hoisting it would
contradict the Apache-2.0 tag on the MLA extract.
